requirements.txt
aes-everywhere==1.2.10
alembic==1.6.5
certifi==2021.10.8
click==8.0.1
cloudinary==1.28.0
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
names==0.3.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycryptodomex==3.10.1
python-dateutil==2.8.2
python-editor==1.0.4
random-username==1.0.2
six==1.16.0
SQLAlchemy==1.4.21
urllib3==1.26.7
Werkzeug==2.0.1
